# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Wedding",
        "description": "Pra casar API for wedding name Pra Casar",
        "termsOfService": "",
        "contact": {
            "email": "admin@approx.com.br"
        },
        "license": {
            "name": "Pro Produtores",
            "url": "https://www.proprodutores.com.br/"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/tiepiece": {
            "get": {
                "tags": [
                    "TiePiece"
                ],
                "summary": "List of TiePieces",
                "description": "Return the List of Tipieces",
                "operationId": "App\\Http\\Controllers\\TiepieceController::index",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TiePiece"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/tiepieces": {
            "post": {
                "tags": [
                    "TiePiece"
                ],
                "summary": "Create TiePiece",
                "description": "Create a Piepiece for the Guests",
                "operationId": "App\\Http\\Controllers\\TiepieceController::create",
                "requestBody": {
                    "description": "TiePiece attributes",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TiePiece"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "List users",
                "description": "Returns the list of all users",
                "externalDocs": "",
                "operationId": "getPetById",
                "parameters": "",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "List users",
                "description": "Returns the list of all users",
                "operationId": "getPetById",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplier"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "TiePiece": {
                "required": [
                    "id",
                    "value",
                    "user_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "value": {
                        "type": "double",
                        "example": 20.63
                    },
                    "user_id": {
                        "type": "string",
                        "example": "TRU87856"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}